requirements.txt
python-dotenv
requests
pexels_api
Flask==2.0.1
